safe_reload.c implements safe reload of haproxy. Invoke with the following arguments:
1:	HAProxy PID file (run time pid file, or a new file to store pid)
2:	HAproxy executable path
3:	List of vip,port,tag as arguments, E.g.:
		10.47.8.252:80:FD_HOST1,10.47.8.252:443:FD_HOST2
4-n:	haproxy arguments (no -sf or -p options, safe_reload will add itself
		using the pid file). E.g.:
	./safe_reload /var/run/ha1/pid /usr/sbin/haproxy \
		10.1.1.1:80:FD_HOST1,10.1.1.2:443:FD_HOST2
		-f /etc/haproxy/haproxy_global.cfg
		-f /etc/haproxy/haproxy_frontend.cfg
		-f /etc/haproxy/haproxy_backend.cfg

Using this progra, the HAProxy configuration file is modified as follows
in the frontend:

	frontend fk-safe-reload
		bind "fd@${FD_HOST1}"
		...

	frontend fk-safe-reload-ssl
		bind "fd@${FD_HOST2}" ssl crt /etc/ssl/haproxy/ssl-file
		...
Where safe_reload creates sockets and binds to the IP addresses as specified
in it's arguments, and sets environment variables as specified in the
'tag's, and invokes HAProxy with the configuration file specified.

To reload the configuration, do the following steps:

1. Make required modifications to the configuration file.
2. Find the process id of safe_reload program - say 'P'
3. Run "kill -USR1 $P

Step #2 and #3 can be merged with:
pkill -USR1 safe_reload

The following script implements the same:

pid=`pgrep safe_reload`
if [ -z $pid ]
then
        exit 1
fi

while :
do
        kill -USR1 $pid
        sleep 0.5
done

TODO:
	- Test results (before and after).
	- Improve option/arguments passing.
	- Other things to implement.

Happy HAPRoxy safe reloading!